+++ changelog.md
@@ -8,6 +8,7 @@
 - Fix for calendar sync on Android permission error on first try
 - Fix Intro page loading container
 - Fix dust font on main page
+- Fix for when you have never opened an event but it has started and you are on cellular
 
 ## 2.22